+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change log html-tag-js
 
+## 2.0.0
+
+- Removed `tag.use` method instead use `html-tag-js/Reactive` to create reactive node. E.g. `const count = Reactive(0);`
+- Now you can pass `onref` callback to `Ref(onrefCallback)` constructor to get reference of the node.
+- Supports promise in children. E.g. `<div>{Promise.resolve('hello')}</div` will render `<div>hello</div>`
+
 ## 1.4.0
 
 - Filter out null and undefined node from children